# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartSchool Asset API",
        "description": "Dokumentasi API untuk sistem manajemen aset SmartSchool. Mencakup peminjaman barang, pengambilan konsumsi, barang masuk, dan laporan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/items": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Daftar semua barang (admin)",
                "description": "Menampilkan daftar semua barang dengan filter nama dan jenis barang. Hanya untuk admin.",
                "operationId": "38f04083708efa13b19c689b47445d64",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari berdasarkan nama barang",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Laptop"
                        }
                    },
                    {
                        "name": "jenis",
                        "in": "query",
                        "description": "Filter jenis barang",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "semua",
                                "aset",
                                "konsumsi"
                            ],
                            "example": "aset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Halaman daftar barang admin"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Barang"
                ],
                "summary": "Simpan barang baru (admin)",
                "description": "Menyimpan data barang baru ke sistem. Mendukung upload foto barang. Hanya untuk admin.",
                "operationId": "5303a71efd2a2fd638dc3ce2bdda9997",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nama_barang",
                                    "kategori",
                                    "jenis_barang",
                                    "stok_awal",
                                    "batas_minimum",
                                    "harga"
                                ],
                                "properties": {
                                    "nama_barang": {
                                        "type": "string",
                                        "example": "Laptop Dell Inspiron"
                                    },
                                    "kategori": {
                                        "type": "string",
                                        "example": "Elektronik"
                                    },
                                    "jenis_barang": {
                                        "type": "string",
                                        "enum": [
                                            "aset",
                                            "konsumsi"
                                        ],
                                        "example": "aset"
                                    },
                                    "merk": {
                                        "type": "string",
                                        "example": "Dell"
                                    },
                                    "hasil_perolehan": {
                                        "type": "string",
                                        "enum": [
                                            "pembelian",
                                            "hibah",
                                            "sumbangan",
                                            "dana_bos"
                                        ],
                                        "example": "pembelian"
                                    },
                                    "id_ruangan": {
                                        "description": "ID ruangan (opsional, untuk filter lemari di form)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_lemari": {
                                        "description": "ID lemari tempat barang disimpan",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "stok_awal": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "batas_minimum": {
                                        "description": "Batas minimum stok sebelum dianggap kritis",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "harga": {
                                        "description": "Harga satuan barang dalam rupiah",
                                        "type": "integer",
                                        "example": 8000000
                                    },
                                    "foto": {
                                        "description": "Foto barang (jpg/jpeg/png, max 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect ke daftar barang jika berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items/export": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Export data barang ke Excel",
                "description": "Download file Excel berisi data semua barang. Mendukung filter yang sama dengan halaman daftar. Hanya untuk admin.",
                "operationId": "727f877fb007a47dd7c716288e252009",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "jenis",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "semua",
                                "aset",
                                "konsumsi"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File Excel (.xlsx)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items/create": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Form tambah barang baru (admin)",
                "description": "Menampilkan form untuk menambahkan barang baru ke sistem. Hanya untuk admin.",
                "operationId": "47ce0f40520af527045dc3143efa2e72",
                "responses": {
                    "200": {
                        "description": "Halaman form tambah barang"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items/{id}/edit": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Form edit barang (admin)",
                "description": "Menampilkan form untuk mengedit data barang beserta riwayat kondisi barang. Hanya untuk admin.",
                "operationId": "a0dcdee17c961aa6677da593b640aee8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID barang",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Halaman form edit barang"
                    },
                    "404": {
                        "description": "Barang tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items/{id}": {
            "put": {
                "tags": [
                    "Barang"
                ],
                "summary": "Update data barang (admin)",
                "description": "Mengupdate data barang yang sudah ada. Foto lama tetap dipertahankan jika tidak ada foto baru yang diupload. Hanya untuk admin.",
                "operationId": "4a2d683f0a5edace6de8e8d0d12c2ee9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID barang",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nama_barang",
                                    "kategori",
                                    "jenis_barang",
                                    "batas_minimum",
                                    "harga"
                                ],
                                "properties": {
                                    "nama_barang": {
                                        "type": "string",
                                        "example": "Laptop Dell Inspiron"
                                    },
                                    "kategori": {
                                        "type": "string",
                                        "example": "Elektronik"
                                    },
                                    "jenis_barang": {
                                        "type": "string",
                                        "enum": [
                                            "aset",
                                            "konsumsi"
                                        ],
                                        "example": "aset"
                                    },
                                    "id_lemari": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "batas_minimum": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "harga": {
                                        "type": "integer",
                                        "example": 8000000
                                    },
                                    "kondisi": {
                                        "type": "string",
                                        "enum": [
                                            "Baik",
                                            "Rusak Ringan",
                                            "Rusak Sedang",
                                            "Rusak Berat",
                                            "Mati Total"
                                        ],
                                        "example": "Baik"
                                    },
                                    "foto": {
                                        "description": "Foto baru (opsional)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect ke daftar barang jika berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "404": {
                        "description": "Barang tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Barang"
                ],
                "summary": "Hapus barang (admin)",
                "description": "Menghapus data barang dari sistem. Hanya untuk admin.",
                "operationId": "f7d8019e7a58101f594ab509931e3bb2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID barang yang akan dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect dengan pesan sukses"
                    },
                    "404": {
                        "description": "Barang tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items/{id}/qr": {
            "get": {
                "tags": [
                    "Barang"
                ],
                "summary": "Download QR Code barang (admin)",
                "description": "Menampilkan halaman cetak QR Code untuk barang tertentu. Hanya untuk admin.",
                "operationId": "2011274e464284d73da489ca0b308f93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID barang",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Halaman cetak QR Code"
                    },
                    "404": {
                        "description": "Barang tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token autentikasi Laravel Sanctum/Session",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}